+++ Source/GUI/preferences.h
@@ -28,8 +28,8 @@ public:
     activealltracks ActiveAllTracks;
     QMap<QString, QString> ActivePanels;
 
-    QList<std::tuple<size_t, size_t>> loadFilterSelectorsOrder();
-    void saveFilterSelectorsOrder(const QList<std::tuple<size_t, size_t>>& order);
+    QList<std::tuple<quint64, quint64>> loadFilterSelectorsOrder();
+    void saveFilterSelectorsOrder(const QList<std::tuple<quint64, quint64>>& order);
 
     bool isSignalServerEnabled() const;
     bool isSignalServerAutoUploadEnabled() const;
